Move clusterstate to admin
