+++ client/go/internal/admin/clusterstate/detect_model.go
