+++ client/go/internal/admin/clusterstate/show_hidden.go
